testing.vmcore: split out management and simplify
The testing.vmcore system requires the OCI SDK to be installed, which
has its own dependencies. Mostly, this is due to the management system
relying on OCI Object Storage, but also we require the "rich" library
for progress bars.

It really doesn't make sense to have extra third-party dependencies
required for testing. They add runtime during CI, there are security
questions, and it is more confusing for developers. Further, using the
OCI SDK for Object Storage requires a full set of OCI API keys, which is
not really suitable for shared systems.

So this commit does two things: first, it splits testing.vmcore into two
sub-components: testing.vmcore.test and testing.vmcore.manage. The
former actually runs the tests, and the latter manages vmcores. The
second task is to migrate vmcore management from the OCI SDK to the
parlib tool introduced in the previous commit. This drops several
dependencies. Further, we drop the rich dependency, resulting in a very
minimal testing script with no more dependencies than necessary.

Orabug: 37307171
