+++ testing/requirements-vmcore.txt
@@ -1,6 +1,3 @@
 # Requirements for running the vmcore tests.
 # Used in Gitlab CI.
 pytest
-# Packages required to manage (upload & download) vmcores for tests
-oci
-rich
